model_service/requirements.txt
# Deep Learning
torch==2.0.1
transformers==4.31.0
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.0.3

# Web Framework
fastapi==0.101.1
uvicorn==0.23.2
starlette==0.27.0

# Database
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.1

# API
pydantic==2.3.0
python-multipart==0.0.6
aiohttp==3.8.5

# Cache & Queue
redis==4.6.0
celery==5.3.1

# Model Optimization
onnx==1.14.0
onnxruntime==1.15.1
tensorrt==8.6.1

# Monitoring & Metrics
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0

# Storage
boto3==1.28.30
oss2==2.18.1
aiofiles==23.2.1

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
httpx==0.24.1

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Documentation
mkdocs==1.5.2
mkdocs-material==9.2.3 